slowly losing will to live
