+++ Sources/CvPropertyManipulators.h
@@ -23,7 +23,7 @@ public:
 	int getNumSources() const;
 	CvPropertySource* getSource(int index) const;
 	const std::vector<CvPropertySource*>& getSources() const { return m_apSources; }
-	const python::list cyGetSources() const;
+	const boost::python::list cyGetSources() const;
 	int addSource(PropertySourceTypes eType);
 
 	int getNumInteractions() const;
