+++ Sources/CvPythonCityLoader.cpp
@@ -1,4 +1,5 @@
 #include "CvPythonCityLoader.h"
+#include "CyCity.h"
 #include <boost/python/manage_new_object.hpp>
