+++ Sources/SCvInternalGlobals.cpp
@@ -4,7 +4,6 @@
 #include "CvGameCoreDLL.h"
 #include "CvBonusInfo.h"
 #include "CvGlobals.h"
-#include "CvInfos.h"
 
 CvBonusInfo* SCvInternalGlobals::getBonusInfoByType(const std::string& szType)
 {
